+++ src/Castle.Services.Transaction.Tests/FileTransaction_WithManager_Tests.cs
@@ -16,6 +16,7 @@
 
 using System;
 using System.IO;
+using System.Transactions;
 
 using NUnit.Framework;
 
@@ -68,7 +69,7 @@ namespace Castle.Services.Transaction.Tests
         [Test]
         public void TransactionResourcesAreDisposed()
         {
-            var tx = _transactionManager.CreateTransaction(TransactionMode.Requires, IsolationMode.Unspecified);
+            var tx = _transactionManager.CreateTransaction(TransactionMode.Requires, IsolationLevel.Unspecified);
 
             var resource = new ResourceImpl();
             tx.Enlist(resource);
@@ -98,7 +99,7 @@ namespace Castle.Services.Transaction.Tests
 
             Assert.That(currentTransaction, Is.Null);
 
-            var tx = _transactionManager.CreateTransaction(TransactionMode.Requires, IsolationMode.Unspecified);
+            var tx = _transactionManager.CreateTransaction(TransactionMode.Requires, IsolationLevel.Unspecified);
 
             tx.Begin();
 
@@ -109,7 +110,7 @@ namespace Castle.Services.Transaction.Tests
 
             // invocation.Proceed() in Castle.DynamicProxy.IInterceptor.Intercept(IInvocation).
 
-            var childTx = _transactionManager.CreateTransaction(TransactionMode.Requires, IsolationMode.Unspecified);
+            var childTx = _transactionManager.CreateTransaction(TransactionMode.Requires, IsolationLevel.Unspecified);
 
             currentTransaction = _transactionManager.CurrentTransaction;
 
@@ -136,7 +137,7 @@ namespace Castle.Services.Transaction.Tests
             // Now we can dispose the main transaction.
             _transactionManager.Dispose(tx);
 
-            Assert.That(txF.Status, Is.EqualTo(TransactionStatus.Committed));
+            Assert.That(txF.Status, Is.EqualTo(Services.Transaction.TransactionStatus.Committed));
             Assert.That(txF.IsDisposed);
         }
 
@@ -149,7 +150,7 @@ namespace Castle.Services.Transaction.Tests
         [Test]
         public void BugWhenResourceFailsAndTransactionCommits()
         {
-            _ = _transactionManager.CreateTransaction(TransactionMode.Requires, IsolationMode.Unspecified);
+            _ = _transactionManager.CreateTransaction(TransactionMode.Requires, IsolationLevel.Unspecified);
         }
     }
 }